gcp_migration_ata/requirements.txt
fastapi
python-multipart
boto3
google-cloud-storage
websocket-client
requests
functions-framework==3.* 